relocate and rename 3
