+++ lib/data/repositories/place_repository.dart
@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
 
 class PlaceRepository {
   PlaceRepository() {
