+++ lib/main.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:places/ui/my_app/my_app.dart';
-
 import 'package:places/ui/my_app/my_app_model.dart';
 import 'package:places/ui/screens/main_screen/main_screen_model.dart';
 import 'package:places/ui/screens/sight_search_screen/sight_search_model.dart';
