+++ lib/ui/my_app/my_app.dart
@@ -1,6 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import 'package:places/ui/my_app/themes.dart';
+import 'package:places/ui/my_app/routes.dart';
+import 'package:places/ui/my_app/platform_detector.dart';
+import 'package:places/ui/my_app/my_app_model.dart';
 import 'package:places/ui/screens/onboarding_screen/onboarding_screen.dart';
 import 'package:places/ui/screens/filter_screen/filters_screen.dart';
 import 'package:places/ui/screens/add_sight_screen/add_sight_screen.dart';
@@ -8,10 +12,6 @@ import 'package:places/ui/screens/select_category_screen/select_category_screen.
 import 'package:places/ui/screens/sight_search_screen/sight_search_screen.dart';
 import 'package:places/ui/screens/main_screen/main_screen.dart';
 import 'package:places/ui/screens/splash_screen/splash_screen.dart';
-import 'package:places/ui/my_app/themes.dart';
-import 'package:places/ui/my_app/routes.dart';
-import 'package:places/ui/my_app/platform.dart';
-import 'package:places/ui/my_app/my_app_model.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
