+++ lib/ui/my_app/platform_detector.dart
