+++ lib/ui/screens/add_sight_screen/add_sight_model.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
-import 'package:places/data/repositories/mocks.dart';
 
+import 'package:places/data/repositories/mocks.dart';
 import 'package:places/data/interactors/repository.dart';
 import 'package:places/data/models/sight.dart';
