+++ lib/ui/screens/add_sight_screen/dialog_add_photo.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:places/ui/res/ui_image_paths.dart';
-import 'package:places/ui/res/ui_strings.dart';
+
+import 'package:places/ui/my_app/ui_image_paths.dart';
+import 'package:places/ui/my_app/ui_strings.dart';
 
 class DialogAddPhoto extends StatelessWidget {
   const DialogAddPhoto({Key? key}) : super(key: key);
