+++ lib/ui/screens/filter_screen/filters_model.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'package:places/data/interactors/repository.dart';
 import 'package:places/data/models/filter_item.dart';
