+++ lib/ui/screens/main_screen/main_1_widget_app_bar.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/res/ui_strings.dart';
+
+import 'package:places/ui/my_app/ui_strings.dart';
 import 'package:places/ui/widgets/headers.dart';
 
 class MySliverAppBar extends SliverPersistentHeaderDelegate {
