+++ lib/ui/screens/main_screen/main_3_0_visiting_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui/widgets/headers.dart';
 import 'package:places/ui/screens/main_screen/main_3_1_tab_wished.dart';
 import 'package:places/ui/screens/main_screen/main_3_2_tab_seen.dart';
+import 'package:places/ui/widgets/headers.dart';
 
 class VisitingScreen extends StatefulWidget {
   const VisitingScreen({Key? key}) : super(key: key);
