+++ lib/ui/screens/main_screen/main_3_1_tab_wished.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:places/ui/my_app/platform.dart';
-import 'package:places/ui/screens/main_screen/widget_add_to_calendar_cuper_modal.dart';
 import 'package:provider/provider.dart';
 
-import 'package:places/data/models/sight.dart';
-
-import 'package:places/ui/my_enums.dart';
-import 'package:places/ui/my_scroll_physics.dart';
+import 'package:places/enums.dart';
+import 'package:places/ui/my_app/my_scroll_physics.dart';
+import 'package:places/ui/my_app/platform_detector.dart';
 import 'package:places/ui/widgets/sight_card.dart';
-
 import 'package:places/ui/screens/main_screen/main_3_3_empty_list.dart';
 import 'package:places/ui/screens/main_screen/main_screen_model.dart';
+import 'package:places/ui/screens/main_screen/widget_add_to_calendar_cuper_modal.dart';
 import 'package:places/ui/screens/sight_details_screen/sight_details_screen.dart';
+
+import 'package:places/data/models/sight.dart';
 
 class TabWished extends StatefulWidget {
   const TabWished({Key? key}) : super(key: key);
