+++ lib/ui/screens/main_screen/main_3_3_empty_list.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/res/text_styles.dart';
+
+import 'package:places/ui/my_app/text_styles.dart';
 
 class WidgetEmptyList extends StatelessWidget {
   const WidgetEmptyList({
