+++ lib/ui/screens/main_screen/main_4_settings_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+
 import 'package:places/ui/my_app/my_app_model.dart';
 import 'package:places/ui/my_app/routes.dart';
-import 'package:places/ui/res/ui_image_paths.dart';
+import 'package:places/ui/my_app/ui_image_paths.dart';
 
 class SettingsScreen extends StatefulWidget {
   const SettingsScreen({Key? key}) : super(key: key);
