+++ lib/ui/screens/onboarding_screen/onboarding_page_2.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui/res/ui_image_paths.dart';
+import 'package:places/ui/my_app/ui_image_paths.dart';
 
 class OnboardingPage2 extends StatelessWidget {
   const OnboardingPage2({
