+++ lib/ui/screens/onboarding_screen/onboarding_page_3.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui/res/ui_image_paths.dart';
+import 'package:places/ui/my_app/ui_image_paths.dart';
 
 class OnboardingPage3 extends StatelessWidget {
   const OnboardingPage3({
