+++ lib/ui/screens/onboarding_screen/onboarding_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:places/ui/res/ui_image_paths.dart';
+import 'package:places/ui/my_app/ui_image_paths.dart';
 import 'package:places/ui/my_app/routes.dart';
-
 import 'package:places/ui/screens/onboarding_screen/onboarding_page_1.dart';
 import 'package:places/ui/screens/onboarding_screen/onboarding_page_2.dart';
 import 'package:places/ui/screens/onboarding_screen/onboarding_page_3.dart';
