+++ lib/ui/screens/sight_details_screen/sight_details_model.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'package:places/data/interactors/repository.dart';
 import 'package:places/data/models/sight.dart';
