+++ lib/ui/screens/sight_details_screen/sight_details_screen.dart
@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:places/data/models/sight.dart';
-import 'package:places/ui/widgets/my_image_widget.dart';
-import 'package:places/ui/res/ui_strings.dart';
-import 'package:places/ui/my_scroll_physics.dart';
+import 'package:places/ui/my_app/ui_strings.dart';
+import 'package:places/ui/my_app/my_scroll_physics.dart';
 import 'package:places/ui/screens/sight_details_screen/sight_details_model.dart';
 import 'package:places/ui/screens/sight_details_screen/widget_back_button.dart';
+import 'package:places/ui/widgets/my_image_widget.dart';
+
+import 'package:places/data/models/sight.dart';
 
 class SightDetailsScreen extends StatefulWidget {
   const SightDetailsScreen({
