+++ lib/ui/screens/splash_screen/splash_model.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
-
 import 'package:flutter/material.dart';
+
 import 'package:places/ui/my_app/routes.dart';
 
 // TODO - avoid_print
