+++ lib/ui/screens/splash_screen/splash_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/res/ui_image_paths.dart';
 
+import 'package:places/ui/my_app/ui_image_paths.dart';
 import 'package:places/ui/screens/splash_screen/splash_model.dart';
 
 class SplashScreen extends StatefulWidget {
