+++ lib/ui/widgets/sight_card.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/models/sight.dart';
-import 'package:places/ui/my_enums.dart';
+
+import 'package:places/enums.dart';
 import 'package:places/ui/widgets/my_image_widget.dart';
 import 'package:places/ui/widgets/sight_card_icons.dart';
+
+import 'package:places/data/models/sight.dart';
 
 class SightCard extends StatelessWidget {
   const SightCard({
