+++ lib/ui/widgets/sight_card_for_search.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+
 import 'package:places/ui/screens/sight_details_screen/sight_details_screen.dart';
-import 'package:places/data/models/sight.dart';
 import 'package:places/ui/widgets/my_image_widget.dart';
+
+import 'package:places/data/models/sight.dart';
 
 class SightCartForSearch extends StatelessWidget {
   const SightCartForSearch({
